extract text customization tab into its own component
